requirements.txt
pygame==2.5.2
google-genai==0.3.0
